+++ javaScript/projects/weather-app/src/components/tabs/today/today.js
@@ -1,0 +1,11 @@
+export default {
+  cacheDOM() {
+    console.log('cacheDOM() running from today.js');
+  },
+  bindEvents() {
+    console.log('bindEvents() running from today.js');
+  },
+  render() {
+    console.log('render() running from today.js');
+  },
+};